Enabling per-player camera following
Also bringing the player back to spawn if they fall off the stage.

Cameras are children of their players and not enabled unless that player
is controlled by the local client, so that lets us have cameras follow
players individually as well as saves us the trouble of writing camera
follow scripts.