Added tiny demo level.
Tiny Level Designed. Here's the idea: The players must stack to boost a
player to position one. The other will then drop to Two. Both circles
must be triggered to open 'the gate' which is the cylinder at the end.

None of that is implemented, but it's a decent demo of what the game
looks like without powers.